Rename modes: pot --> update-sources, po --> update-translations
The new names are more precise and easier to understand.
