+++ checkmk_weblate_syncer/__main__.py
@@ -3,10 +3,10 @@ from pathlib import Path
 from typing import TypeVar, assert_never
 
 from .cli import Mode, parse_arguments
-from .config import PoModeConfig, PotModeConfig
+from .config import UpdateSourcesConfig, UpdateTranslationsConfig
 from .logging import LOGGER, configure_logger
-from .po import run as run_po_mode
-from .pot import run as run_pot_mode
+from .update_sources import run as run_update_sources
+from .update_translations import run as run_update_translations
 
 
 def _main() -> None:
@@ -14,15 +14,21 @@ def _main() -> None:
     configure_logger(args.log_level)
 
     match args.mode:
-        case Mode.POT:
-            sys.exit(run_pot_mode(_load_config(args.config_path, PotModeConfig)))
-        case Mode.PO:
-            sys.exit(run_po_mode(_load_config(args.config_path, PoModeConfig)))
+        case Mode.UPDATE_SOURCES:
+            sys.exit(
+                run_update_sources(_load_config(args.config_path, UpdateSourcesConfig))
+            )
+        case Mode.UPDATE_TRANSLATIONS:
+            sys.exit(
+                run_update_translations(
+                    _load_config(args.config_path, UpdateTranslationsConfig)
+                )
+            )
         case _:
             assert_never(args.mode)
 
 
-_ConfigTypeT = TypeVar("_ConfigTypeT", PotModeConfig, PoModeConfig)
+_ConfigTypeT = TypeVar("_ConfigTypeT", UpdateSourcesConfig, UpdateTranslationsConfig)
 
 
 def _load_config(config_path: Path, config_type: type[_ConfigTypeT]) -> _ConfigTypeT:
